pyproject.toml
[project]
name = "ivp-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "psycopg2>=2.9.11",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "reportlab>=4.4.7",
    "svglib>=1.6.0",
    "uvicorn>=0.40.0",
]
